device: explicitely setup the full path to the cross_compile toolchains.
this is to work around cases where the PATH global shell environment
is not exported because the 'setpath' is not invoked (when not starting
a build through the 'lunch' combo command as example).

this is arguably a bit hacky, but we need this for the time being to
pass the compiler properly to the sub-make's invoked for refsw.  eventually
we will remove all this cruft and not need this setting...

refs #SWANDROID-2346

Change-Id: Iae4f037c8f9b71967c90b5357d85e80b335cfcea
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>