major update to "bcm_refsw.mk"
- need to avoid a race condition caused by calling the top-level Makefile
  recursively

When performing sub-make's, do not call the top-level Makefile with
a specified target, as this causes all of the "Android.mk" files to
be rescanned.

- improve handling of targets built by Android.mk rules and those built
  by custom "makefile" rules

Previously, the "refsw" target, defined in this makefile, built targets
required by "libnexuseglclient" (an Android.mk target), then called the
top-level Makefile (to build "libnexuseglclient"), and then built the
targets which required this "libnexuseglclient" library (with these custom
"makefile" rules). By eliminating the need to call the top-level Makefile,
the requirement to explicitly build "libnexuseglclient" also is eliminated.
As a result, the targets which were previously triggering the "refsw_build"
can now trigger either "nexus_build" or "brcm_dhd_driver" appropriately,
and the "v3d_driver" can be an appropriate dependency.

- cleanup the .PHONY rules
- add additional logging messages

refs #SWANDROID-2399

Change-Id: I8fd19ffd8967f43eb21f2ecc37c69830efa64830
Signed-off-by: Steve Rae <srae@broadcom.com>