GPT: Remove unnecessary "gpt" entry in the partition table.
The fake "gpt" entry in the partition table was previously used to
allow fastboot to flash a new gpt.bin at offset 0. Technically, this
is an invalid GPT entry as no partition should start before the
"first usable LBA" (34) and some GPT tools and kernels complain about
that (no big deal, it still worked).

Our Bolt fastboot code understands the name "gpt" as a special case
for offset 0 and doesn't need the partition table to contain a fake
"gpt" entry.

refs #SWANDROID-2558

Change-Id: I949a48fe4ace969b8fcce761a1f15e8d4928cc21