config: for platform with encoder, increase nexus main heap size.
some of the vce internal buffers have been moved from crr region to
glr (open) which means effectively into main heap 0 (see SWSTB-2083).

consequently, the heap 0 usage on the android platform has increased
due to this ~5.5MB swap in heap assignment and has caused heap 0 usage
to go beyond its maximum target for worst case usage (thus failing).

in order to continue to support the needed features, we must therefore
increase heap 0.  at this time we continue to require 16MB alignment
across all heaps (this will change in 16.3), therefore we make the
heap 0 increase from 96MB to 112MB to absorb the ~6MB demand from vce
and keep the 16MB alignment across nexus.

the crr heap could be reduced in consequence (since it is now ~6MB less
used); however because of the alignment requirement, it would end up
being too small for the usage if taken to the lower notch size, so we
keep it where it is for now.

in the 16.3 timeframe, we can refine the tuning of the heaps once we
no longer have the 16MB alignment restriction.

updated boot command line in consequence for platforms with increased
heap 0.

refs #SWANDROID-3338

Change-Id: I501d45bc575fc72e9b5787766d1efe299088cc61
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>