avko: increase the lmk threshold.
by increasing the lmk threshold, we want to force memory reclaim
to kick in earlier which helps with scenarios where a lots of applications
are open one after the other continuously leading to device becoming
sluggish as reclaim/swap takes place.  by increasing the limit of the lmk
we allow more background killing to improve the reclaim/swap time and
need; overall giving a better user experience.

b/29521215
refs #SWANDROID-3449

Change-Id: I57d1e2e9fccac64dd0b06c0e35f657915dd37e69
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>